backend/pyproject.toml
[project]
name = "news-neutralizer-api"
version = "0.1.0"
description = "News Neutralizer — Backend API for news analysis and bias detection"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pgvector>=0.2.0",
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "gevent>=24.2.1",
    "psycogreen>=1.0.2",
    # HTTP Client (async)
    "httpx>=0.28.0",
    # Scraping
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "feedparser>=6.0.0",
    "duckduckgo-search>=6.0.0",
    "googlenewsdecoder>=0.1.0",
    # NLP / Embeddings (local)
    "sentence-transformers>=3.3.0",
    "scikit-learn>=1.4.0",
    "lingua-language-detector>=2.0.0",
    # AI Providers
    "openai>=1.58.0",
    "anthropic>=0.42.0",
    "google-generativeai>=0.8.0",
    # Security
    "cryptography>=44.0.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    # Logging
    "structlog>=24.4.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Asyncio helpers
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
